transcribe-ws/requirements.txt
Flask==2.3.2
openai-whisper==20231117
torch==2.1.0
gunicorn==21.2.0